Rakefile: Change git push to force
Force the git push on _deploy stuffs.
This is because I use Octopress on different machines,
and everytime I generate a new website, I just wanna
push it to github anyway, it doesn't make sense to
keep the history of _deploy stuffs.

Maybe this should be submitted to Octopress as well?

Signed-off-by: Mark Zhang <markz@nvidia.com>