HACK: Recognize .h file as cpp
include_code.rb has a bug that it can't recognize .h file.
It reports "h" to pygments so it triggers an exception of
pygments to say that "h" is an unknown language.

So this is a HACK to convert "h" -> "cpp". This makes we
can "include_code" a .h file.

Signed-off-by: Mark Zhang <super119@139.com>